split out cap_max_gas_price tests
