+++ integration-tests/src/tests/client/features.rs
@@ -2,6 +2,7 @@
 
 #[cfg(feature = "protocol_feature_account_id_in_function_call_permission")]
 mod account_id_in_function_call_permission;
+mod cap_max_gas_price;
 #[cfg(feature = "protocol_feature_fix_contract_loading_cost")]
 mod fix_contract_loading_cost;
 mod fix_storage_usage;
